Added code to send a URB and receive a result every 1HZ.
This will eventually be able to unblock read calls on portC. It does not send useful data right now (it sends a kzalloced block of memory. Eventually it will to an AIN read. Unfortunately, the access to the USB device is not locked yet, so the order in which IN URBs are returned could be mussed up.